refactor: move Hardware views to view.py
